v0.2.13: load subcommand filters, keep cargo Finished line, fix clippy
- Load <base>-*.toml filters too (git-log.toml, git-diff.toml, cargo-tree.toml
  were never being applied — only <base>.toml loaded). Big quality win on
  git log SHA truncation, git diff context collapse, etc.
- Keep `Finished` and `Building` cargo lines so no-op builds still confirm
  success (was producing zero output, leaving users in the dark).
- Fix clippy unnecessary_cast warning in tkr-server.

+++ crates/tkr-server/src/main.rs
@@ -392,7 +392,7 @@ async fn write_event_triplet(io: &mut TokioIo<Upgraded>, state: &AppState) -> an
         "type": "command_end",
         "cmd": cmd,
         "tokensIn": 1800,
-        "tokensSaved": 800 + ((next_event_id(state) % 3200) as u64),
+        "tokensSaved": 800 + (next_event_id(state) % 3200),
     });
 
     for event in [start, line, end] {
